calculate using doubles

+++ cpp09/ex01/RPN.cpp
@@ -36,7 +36,7 @@ void RPN::push_val(std::string &tok) {
 void RPN::_plus() {
   if (_stack.size() < 2)
     throw std::exception();
-  int val = _stack.top();
+  double val = _stack.top();
   _stack.pop();
   val += _stack.top();
   _stack.pop();
@@ -46,9 +46,9 @@ void RPN::_plus() {
 void RPN::_minus() {
   if (_stack.size() < 2)
     throw std::exception();
-  int val1 = _stack.top();
+  double val1 = _stack.top();
   _stack.pop();
-  int val2 = _stack.top();
+  double val2 = _stack.top();
   _stack.pop();
   _stack.push(val2 - val1);
 }
@@ -56,7 +56,7 @@ void RPN::_minus() {
 void RPN::_multiply() {
   if (_stack.size() < 2)
     throw std::exception();
-  int val = _stack.top();
+  double val = _stack.top();
   _stack.pop();
   val *= _stack.top();
   _stack.pop();
@@ -66,14 +66,14 @@ void RPN::_multiply() {
 void RPN::_divide() {
   if (_stack.size() < 2)
     throw std::exception();
-  int val1 = _stack.top();
+  double val1 = _stack.top();
   _stack.pop();
-  int val2 = _stack.top();
+  double val2 = _stack.top();
   _stack.pop();
   _stack.push(val2 / val1);
 }
 
-int RPN::calculate(const std::string &arg) {
+double RPN::calculate(const std::string &arg) {
   RPN rpn;
   std::string token;
   std::istringstream iss(arg);
